Use Long for ID's instead of Int's
Room DB uses Long for number ID's.

+++ app/src/main/java/net/hawkins/gamescore/data/model/FavoriteGame.kt
@@ -4,5 +4,5 @@ data class FavoriteGame(
     val name: String = "",
     val players: List<String> = emptyList(),
     val game: Game,
-    override var id: Int = 0
+    override var id: Long = 0
 ) : Idable
