+++ app/src/main/java/net/hawkins/gamescore/ui/managegames/GameManagementUiEvent.kt
@@ -4,6 +4,6 @@ import androidx.compose.runtime.Stable
 
 @Stable
 sealed interface  GameManagementUiEvent {
-    data class DeleteGame(val id: Int) : GameManagementUiEvent
+    data class DeleteGame(val id: Long) : GameManagementUiEvent
     object RefreshState : GameManagementUiEvent
 }
